feat: added semantic viewcourse.html.erb with typography.scss. fix: used rails g controller courses instead of course to create files, removed old files
