+++ app/assets/stylesheets/courses.css
@@ -1,0 +1,1 @@
+@import "typography";
