+++ app/helpers/courses_helper.rb
@@ -1,0 +1,2 @@
+module CoursesHelper
+end
